requirements.txt
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
google-auth==2.11.0
google-auth-oauthlib==0.5.3
gspread==5.5.0
httplib2==0.20.4
idna==3.4
libusb1==3.0.0
ndeflib==0.3.3
nfcpy==1.0.4
oauth2client==4.1.3
oauthlib==3.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDes==2.0.1
pyparsing==3.0.9
pyserial==3.5
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==1.26.12
